Elixir2pdf v0.0.1 Elixir2pdf

Elixir2pdf is a simple Elixir wrapper for ErlGuten, a Erlang PDF generation application. See README.md for a general overview.

Functions

center_aligned_text(pid, x1, x2, y, set_font, text)

Adds text to a PDF process that is centered between the two horizontal coordinate points given.

Parameters

pid - the process identifier of the current PDF process.

x1 - the starting horizontal coordinate point in integer format.

x2 - the ending horizontal coordinate point in integer format.

y - the vertical coordinate point in integer format.

set_font - a tuple containing the name and font size of the current font set (i.e. {“Helvetica”, 8}).

text - the actual text to be added in string format.

close(pid)

Deletes the PDF process at the given identifier.

Parameters

pid - the process identifier of the current PDF process.

draw_line(pid, from, to, width, color \\ :black)

Draws a line between the two coordinate points given.

Parameters

pid - the process identifier of the current PDF process.

from - a tuple containing the x and y coordinate points of the start of a line (i.e. {0, 0}).

to - a tuple containing the x and y coordinate points of the end of a line (i.e. {72, 72}).

width - the desired thickness of the line in integer or float format.

color - name of line color in atom format. See the COLORS.md documentation for a list of available colors. Default is :black.

draw_rectangle(pid, basepoint, size, width, color \\ :black)

Draws a rectangle starting at a given basepoint of a given size.

Parameters

pid - the process identifier of the current PDF process.

basepoint - a tuple containing the coordinate points where the bottom-left corder of the rectangle will be located (i.e. {0, 0}).

size - a tuple containing the width and height dimensions of the image (i.e. {64, 89}).

width - the desired thickness of the line in integer or float format.

color - name of line color in atom format. See the COLORS.md documentation for a list of available colors. Default is :black.

export(pid, path)

Exports the PDF process to a PDF file at the given path.

Parameters

pid - the process identifier of the current PDF process.

path - path where the generated PDF file will be saved in string format (i.e. “./example.pdf”).

insert_image(pid, path, basepoint, size)

Inserts an image into the PDF from a given path.

Parameters

pid - the process identifier of the current PDF process.

path - path where the image file is located in string format (i.e. “./example.jpg”).

basepoint - a tuple containing the coordinate points where the bottom-left corder of the image will be located (i.e. {0, 0}).

size - a tuple containing the width and height dimensions of the image (i.e. {64, 89}).

new()

Starts a new PDF process and returns a process identifier (pid).

right_aligned_text(pid, set_font, x, y, text)

Adds text to a PDF process at a given coordinate with right-hand justification.

Parameters

pid - the process identifier of the current PDF process.

set_font - a tuple containing the name and font size of the current font set (i.e. {“Helvetica”, 8}).

x - the horizontal coordinate point where the text will end in integer format.

y - the vertical coordinate point in integer format.

text - the actual text to be added in string format.

set_font(pid, font, color \\ :black)

Sets the font that will apply to all text that follows.

Parameters

pid - the process identifier of the current PDF process.

font - a tuple containing the name of the font in string format and the font size in integer format. See the FONTS.md documentation for a list of available fonts.

color - name of font color in atom format. See the COLORS.md documentation for a list of available colors. Default is :black.

set_new_page(pid)

Creates a new page and returns a process identifier (pid). Any subsequent activity will occur on the new page. Note: Elixir2pdf.new/0 takes care of this, so it is not necessary for the first page.

Parameters

pid - the process identifier of the current PDF process.

text(pid, x, y, text)

Adds text to a PDF process at a given coordinate. See the README.md file for additional detail on how PDF document navigation works.

Parameters

pid - the process identifier of the current PDF process.

x - the horizontal coordinate point in integer format.

y - the vertical coordinate point in integer format.

text - the actual text to be added in string format.